requirements.txt
langchain
langchain_community
langchain-google-genai
python-dotenv
streamlit
langchainhub
faiss-cpu
pypdf